pyproject.toml
[project]
name = "adk-multiagent-systems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audio-recorder-streamlit>=0.0.10",
    "fastapi>=0.115.12",
    "google-adk[extensions]==1.0.0",
    "google-cloud-logging>=3.12.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.2.14",
    "newsapi-python>=0.2.7",
    "newsdataapi>=0.1.23",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "uvicorn>=0.34.2",
    "wikipedia>=1.4.0",
]
